pyproject.toml
[project]
name = "marker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marker-pdf>=1.10.2",
    "pillow>=10.4.0",
    "triton>=3.6.0",
]
